+++ src/Cli/SubCommands/index.ts
@@ -4,6 +4,7 @@ import HelpCommand from "./HelpCommand"
 import Command from "./Command"
 import StartServerCommand from "./StartServerCommand"
 import OpenWebUiCommand from "./OpenWebUiCommand"
+import StopServerCommand from "./StopServerCommand"
 
 const commands: Command[] = [
     GetVersionCommand,
@@ -11,6 +12,7 @@ const commands: Command[] = [
     OpenWebUiCommand,
     OpenUiCommand,
     StartServerCommand,
+    StopServerCommand,
 ]
 
 export default commands